+++ src/ts-main/index.ts
@@ -1,11 +1,14 @@
 import logger, {
   LoggerLevel as innerLoggerLevel,
 } from "../ts-common/logger.ts";
-import type { WaspHlsPlayerConfig } from "../ts-common/types.ts";
+import type {
+  InitialAudioTrackSelection,
+  WaspHlsPlayerConfig,
+} from "../ts-common/types.ts";
 import WaspHlsPlayer from "./api.ts";
 import { PlayerState as innerPlayerState } from "./types.ts";
 
-export type { WaspHlsPlayerConfig };
+export type { InitialAudioTrackSelection, WaspHlsPlayerConfig };
 export type { AudioTrackInfo, VariantInfo } from "./types.ts";
 export type { WaspError } from "./errors/index.ts";
 export {